feat!: revive mapping, state_lattice, and decision modules
The dormant modules return to the build, gate-clean, as the next
development frontier:

- mapping/map_processing: builds against PCL/OpenCV; pnm/rapidcsv are
  implementation-private.
- planning/state_lattice: depends on geometry; drawing sources build
  only with visualization (geometry pattern); the retired FolderPath
  utility is replaced by an env-based data-path shim (TODO: explicit
  configuration).
- planning/decision: markov and reachability build with real
  dependency lists (stale 'control'/'utilities' target names removed);
  GSL is gone (see previous commit). prediction stays excluded — its
  core headers depend on the retired traffic_map module.
- Ghost-dependent apps/tests (traffic_map, trafficviz, file_io) are
  excluded with explicit markers; compiling dev binaries build but are
  not ctest-registered (they run simulations).
- API catch-up: quickviz::CvCanvas/CvIO qualification, stopwatch's
  xmotion namespace + stopwatch/ include path, SystemPropagator's
  single-parameter template, test_matrix's 2018-era comma-initializer
  misuse.
- Export set completed: every aggregate member (now including event,
  math_utils, common types, and the revived modules) installs into
  xmNavigationTargets; vendored header-only deps (stb/pnm/rapidcsv/
  stopwatch) join the export set with exportable include paths.
- libgsl-dev removed from CI and deb dependencies.

Verified: viz-on 23/23, viz-off 14/14, ASan+UBSan 14/14 plus sampler
dev-binary smoke runs.

+++ src/mapping/map_processing/src/pgm_map.cpp
@@ -38,8 +38,8 @@ bool PgmMap::LoadData() {
   auto height = pgm.y_size();
   XM_INFO("PGM image size: {}x{}", width, height);
   data_ = Eigen::MatrixXi::Zero(height, width);
-  for (long y = 0; y < height; ++y) {
-    for (long x = 0; x < width; ++x) {
+  for (long y = 0; y < static_cast<long>(height); ++y) {
+    for (long x = 0; x < static_cast<long>(width); ++x) {
       data_(y, x) = pgm[y][x].value;
     }
   }
